Rename the Spring Boot autoconfigure module to drop MongoDB from its name
Step 2 of #409. The module called occurrent-mongodb-spring-boot-autoconfigure
is already store-neutral: only 2 of its 9 files touch anything Mongo, and
both only through the TimeRepresentation enum, whose own module depends on
nothing but jspecify. The name is wrong, not the contents.

Rename three things together:

  directory   framework/spring-boot-autoconfigure-mongodb-common
           -> framework/spring-boot-autoconfigure/common
  artifactId  occurrent-mongodb-spring-boot-autoconfigure
           -> occurrent-spring-boot-autoconfigure
  package     org.occurrent.springboot.mongo.common
           -> org.occurrent.springboot.common

Add the framework/spring-boot-autoconfigure aggregate POM, which the
per-stack modules will join in the next step, and exclude it from Central
publishing the way the other aggregates are. Update the BOM and both
starter POMs.

The name keeps Spring's third-party <qualifier>-spring-boot-autoconfigure
shape, which is the rule ADR 0055 set when it moved the starters off the
reserved spring-boot-starter- prefix.

This is the second coordinate rename for this artifact, after the 0.30 one
documented in doc/migration/upgrading-to-0.30.0.md, and ADR 0055 warned
against exactly that. Doing it inside a migration users must already
perform for 0.31.0 is what makes it acceptable, and a third would not be.
The OpenRewrite entries and the migration note come in a later step.

PackageMoveRenameTest is deliberately untouched, since it names the old
package as fixture data for the 0.30 recipe.

Verified: mvn -o validate over all 204 modules is BUILD SUCCESS, and the
99 container-free tests across the autoconfigure module and both starters
pass (23 + 56 + 20, zero failures).
