+++ bom/pom.xml
@@ -519,7 +519,7 @@
             </dependency>
             <dependency>
                 <groupId>org.occurrent</groupId>
-                <artifactId>occurrent-mongodb-spring-boot-autoconfigure</artifactId>
+                <artifactId>occurrent-spring-boot-autoconfigure</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
