+++ framework/spring-boot-autoconfigure/common/pom.xml
@@ -19,12 +19,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>org.occurrent</groupId>
-        <artifactId>framework</artifactId>
+        <artifactId>spring-boot-autoconfigure</artifactId>
         <version>${revision}</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>occurrent-mongodb-spring-boot-autoconfigure</artifactId>
+    <artifactId>occurrent-spring-boot-autoconfigure</artifactId>
 
     <name>${mavenProjectName}</name>
