+++ framework/spring-boot-autoconfigure/common/src/main/java/org/occurrent/springboot/common/Jackson3CloudEventConverterConfiguration.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.occurrent.springboot.mongo.common;
+package org.occurrent.springboot.common;
 
 import org.occurrent.application.converter.CloudEventConverter;
 import org.occurrent.application.converter.jackson3.JacksonCloudEventConverter;
