+++ framework/spring-boot-autoconfigure/common/src/main/java/org/occurrent/springboot/common/OnDomainEventQueriesCapabilityCondition.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.occurrent.springboot.mongo.common;
+package org.occurrent.springboot.common;
 
 import org.jspecify.annotations.NonNull;
 import org.occurrent.eventstore.api.EventStoreCapability;
@@ -30,24 +30,12 @@ import java.util.Set;
 import static org.occurrent.eventstore.api.EventStoreCapability.DCB;
 import static org.occurrent.eventstore.api.EventStoreCapability.STREAM;
 
-/**
- * Matches when the auto-configured event store writes a global position. That is the case when DCB is enabled (DCB
- * always writes position) or when STREAM is enabled with {@code occurrent.event-store.stream.position} on (the
- * default). Use it to gate beans that only make sense when the store writes position.
- */
-public class OnPositionEnabledCondition implements Condition {
+public class OnDomainEventQueriesCapabilityCondition implements Condition {
     @Override
     public boolean matches(ConditionContext context, @NonNull AnnotatedTypeMetadata metadata) {
-        Binder binder = Binder.get(context.getEnvironment());
-        Set<EventStoreCapability> capabilities = binder
+        Set<EventStoreCapability> capabilities = Binder.get(context.getEnvironment())
                 .bind("occurrent.event-store.capabilities", Bindable.setOf(EventStoreCapability.class))
                 .orElse(Set.of(STREAM));
-        if (capabilities.contains(DCB)) {
-            return true;
-        }
-        boolean streamPositionEnabled = binder
-                .bind("occurrent.event-store.stream.position", Bindable.of(Boolean.class))
-                .orElse(Boolean.TRUE);
-        return capabilities.contains(STREAM) && streamPositionEnabled;
+        return capabilities.contains(STREAM) || capabilities.contains(DCB);
     }
 }
