+++ framework/spring-boot-autoconfigure/common/src/main/java/org/occurrent/springboot/common/OnStreamEventStoreCapabilityCondition.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.occurrent.springboot.mongo.common;
+package org.occurrent.springboot.common;
 
 import org.jspecify.annotations.NonNull;
 import org.occurrent.eventstore.api.EventStoreCapability;
@@ -27,15 +27,14 @@ import org.springframework.core.type.AnnotatedTypeMetadata;
 
 import java.util.Set;
 
-import static org.occurrent.eventstore.api.EventStoreCapability.DCB;
 import static org.occurrent.eventstore.api.EventStoreCapability.STREAM;
 
-public class OnDomainEventQueriesCapabilityCondition implements Condition {
+public class OnStreamEventStoreCapabilityCondition implements Condition {
     @Override
     public boolean matches(ConditionContext context, @NonNull AnnotatedTypeMetadata metadata) {
         Set<EventStoreCapability> capabilities = Binder.get(context.getEnvironment())
                 .bind("occurrent.event-store.capabilities", Bindable.setOf(EventStoreCapability.class))
                 .orElse(Set.of(STREAM));
-        return capabilities.contains(STREAM) || capabilities.contains(DCB);
+        return capabilities.contains(STREAM);
     }
 }
