+++ framework/spring-boot-autoconfigure/common/src/main/java/org/occurrent/springboot/common/package-info.java
@@ -15,6 +15,6 @@
  */
 
 @NullMarked
-package org.occurrent.springboot.mongo.common;
+package org.occurrent.springboot.common;
 
 import org.jspecify.annotations.NullMarked;
