+++ framework/spring-boot-starter-mongodb-reactive/src/main/java/org/occurrent/springboot/mongo/reactor/OccurrentReactiveMongoAutoConfiguration.java
@@ -43,13 +43,13 @@ import org.occurrent.eventstore.api.reactor.EventStoreQueries;
 import org.occurrent.eventstore.mongodb.spring.reactor.EventStoreConfig;
 import org.occurrent.eventstore.mongodb.spring.reactor.ReactorMongoEventStore;
 import org.occurrent.filter.Filter;
-import org.occurrent.springboot.mongo.common.Jackson3CloudEventConverterConfiguration;
-import org.occurrent.springboot.mongo.common.OccurrentProperties;
-import org.occurrent.springboot.mongo.common.OccurrentProperties.EventStoreProperties;
-import org.occurrent.springboot.mongo.common.OnDcbEventStoreCapabilityCondition;
-import org.occurrent.springboot.mongo.common.OnDomainEventQueriesCapabilityCondition;
-import org.occurrent.springboot.mongo.common.OnMissingCloudEventConverterAndCloudEventTypeMapperCondition;
-import org.occurrent.springboot.mongo.common.OnStreamEventStoreCapabilityCondition;
+import org.occurrent.springboot.common.Jackson3CloudEventConverterConfiguration;
+import org.occurrent.springboot.common.OccurrentProperties;
+import org.occurrent.springboot.common.OccurrentProperties.EventStoreProperties;
+import org.occurrent.springboot.common.OnDcbEventStoreCapabilityCondition;
+import org.occurrent.springboot.common.OnDomainEventQueriesCapabilityCondition;
+import org.occurrent.springboot.common.OnMissingCloudEventConverterAndCloudEventTypeMapperCondition;
+import org.occurrent.springboot.common.OnStreamEventStoreCapabilityCondition;
 import org.occurrent.subscription.api.reactor.CheckpointAwareSubscriptionModel;
 import org.occurrent.subscription.api.reactor.Subscribable;
 import org.occurrent.subscription.api.reactor.SubscriptionModel;
